moved run time warning log to logging client
